tags added + primary Id types changed

+++ src/TimeHacker.Application/Models/Input/Tasks/InputDynamicTaskModel.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string? Description { get; set; }
         public IEnumerable<int> CategoryIds { get; set; }
-        public uint Priority { get; set; }
+        public byte Priority { get; set; }
         public TimeSpan MinTimeToFinish { get; set; }
         public TimeSpan MaxTimeToFinish { get; set; }
         public TimeSpan? OptimalTimeToFinish { get; set; }
